+++ django/applications/catmaid/urls.py
@@ -147,6 +147,7 @@ urlpatterns += [
 # Treenode labels
 urlpatterns += [
     url(r'^(?P<project_id>\d+)/labels/$', label.labels_all),
+    url(r'^(?P<project_id>\d+)/labels/stats$', label.get_label_stats),
     url(r'^(?P<project_id>\d+)/labels-for-nodes$', label.labels_for_nodes),
     url(r'^(?P<project_id>\d+)/labels/(?P<node_type>(treenode|location|connector))/(?P<node_id>\d+)/$', label.labels_for_node),
     url(r'^(?P<project_id>\d+)/label/(?P<ntype>(treenode|location|connector))/(?P<location_id>\d+)/update$', record_view("labels.update")(label.label_update)),
@@ -252,6 +253,7 @@ urlpatterns += [
     url(r'^(?P<project_id>\d+)/skeletons/import$', record_view("skeletons.import")(skeleton.import_skeleton)),
     url(r'^(?P<project_id>\d+)/skeleton/annotationlist$', skeleton.annotation_list),
     url(r'^(?P<project_id>\d+)/skeletons/within-spatial-distance$', skeleton.within_spatial_distance),
+    url(r'^(?P<project_id>\d+)/skeletons/node-labels$', skeleton.skeletons_by_node_labels),
     url(r'^(?P<project_id>\d+)/skeletongroup/adjacency_matrix$', skeleton.adjacency_matrix),
     url(r'^(?P<project_id>\d+)/skeletongroup/skeletonlist_subgraph', skeleton.skeletonlist_subgraph),
     url(r'^(?P<project_id>\d+)/skeletongroup/all_shared_connectors', skeleton.all_shared_connectors),